requirements.txt
matplotlib==3.4.3
muselsl==2.2.2
numpy==1.19.5
playsound==1.3.0
pygame==2.4.0
pylsl==1.10.5
python-dotenv==1.0.0
scikit_learn==1.0.2
scipy==1.7.3
